LayoutTests/platform/wpe/imported/w3c/web-platform-tests/xhr/send-redirect-infinite-expected.txt

CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8800/XMLHttpRequest/resources/infinite-redirects.py?page=default&type=301&mix=0 due to access control checks.

PASS XMLHttpRequest: send() - Redirects (infinite loop) 
